requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.3
django-heroku==0.3.1
future==0.18.2
gunicorn==20.1.0
idna==2.10
kiwisolver==1.3.1
line-bot-sdk==1.19.0
lxml==4.6.3
matplotlib==3.4.2
mpl-finance==0.10.1
mplfinance==0.12.7a17
multitasking==0.0.9
numpy==1.20.3
pandas==1.2.4
pandas-datareader==0.9.0
Pillow==8.2.0
psycopg2==2.8.6
pyimgur==0.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
static3==0.7.0
typing-extensions==3.10.0.0
urllib3==1.26.4
whitenoise==5.2.0
yfinance==0.1.63
